Move helpers to their own modules and add testing framework
