+++ resources/plugin.rb
@@ -57,55 +57,7 @@ action :delete do
 end
 
 action_class do
-  # return the path property if specified or
-  # CHEF_CONFIG_PATH/ohai/plugins if a path isn't specified
-  def desired_plugin_path
-    if new_resource.path
-      new_resource.path
-    else
-      ::File.join(chef_config_path, 'ohai', 'plugins')
-    end
-  end
-
-  # return the chef config files dir or fail hard
-  def chef_config_path
-    if Chef::Config['config_file']
-      ::File.dirname(Chef::Config['config_file'])
-    else
-      Chef::Application.fatal!("No chef config file defined. Are you running \
-chef-solo? If so you will need to define a path for the ohai_plugin as the \
-path cannot be determined")
-    end
-  end
-
-  # is the desired plugin dir in the ohai config plugin dir array?
-  def in_plugin_path?(path)
-    # get the directory where we plan to stick the plugin (not the actual file path)
-    desired_dir = ::File.directory?(path) ? path : ::File.dirname(path)
-
-    case node['platform']
-    when 'windows'
-      ::Ohai::Config.ohai['plugin_path'].map(&:downcase).include?(desired_dir.downcase)
-    else
-      ::Ohai::Config.ohai['plugin_path'].include?(desired_dir)
-    end
-  end
-
-  def add_to_plugin_path(path)
-    ::Ohai::Config.ohai['plugin_path'] << path # new format
-  end
-
-  # we need to warn the user that unless the path for this plugin is in Ohai's
-  # plugin path already we're going to have to reload Ohai on every Chef run.
-  # Ideally in future versions of Ohai /etc/chef/ohai/plugins is in the path.
-  def plugin_path_warning
-    Chef::Log.warn("The Ohai plugin_path does not include #{desired_plugin_path}. \
-Ohai will reload on each chef-client run in order to add this directory to the \
-path unless you modify your client.rb configuration to add this directory to \
-plugin_path. The plugin_path can be set via the chef-client::config recipe. \
-See 'Ohai Settings' at https://docs.chef.io/config_rb_client.html#ohai-settings \
-for more details.")
-  end
+  include OhaiCookbook::PluginHelpers
 end
 
 # this resource forces itself to run at compile_time
