+++ spec/unit/libraries/hint_helper_spec.rb
@@ -1,0 +1,9 @@
+require 'spec_helper'
+
+require_relative '../../../libraries/hint_helpers.rb'
+
+describe OhaiCookbook::HintHelpers do
+  let(:helper) do
+    Object.new.extend(OhaiCookbook::HintHelpers)
+  end
+end
